splitting out the transformations for doing something with (ie. currently inlining all) conjugations (wip - there are fixmes in the code)
